+++ ipc/socket/vsock/vsock_helpers.c
@@ -1,0 +1,28 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <fcntl.h>
+#include <unistd.h>
+#include <sys/ioctl.h>
+#include <sys/socket.h>
+#include <linux/vm_sockets.h>
+
+#include "vsock_helpers.h"
+
+
+int vsock_get_local_cid(void)
+{
+	unsigned int cid;
+	int fd;
+
+	fd = open(DEV_VSOCK, O_RDONLY);
+	if (fd == -1) {
+		fprintf(stderr, "open(%s) failed, %m\n", DEV_VSOCK);
+		return -1;
+	}
+
+	ioctl(fd, IOCTL_VM_SOCKETS_GET_LOCAL_CID, &cid);
+
+	close(fd);
+	return cid;
+}