+++ ipc/socket/vsock/vsock_helpers.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define DEV_VSOCK	"/dev/vsock"
+
+int vsock_get_local_cid(void);